+++ 02-express-tutorial/data.js
@@ -31,12 +31,12 @@ const products = [
     price: 9.99,
     desc: `I'm baby direct trade farm-to-table hell of, YOLO readymade raw denim venmo whatever organic gluten-free kitsch schlitz irony af flexitarian.`,
   },
-];
+]
 const people = [
   { id: 1, name: 'john' },
   { id: 2, name: 'peter' },
   { id: 3, name: 'susan' },
   { id: 4, name: 'anna' },
   { id: 5, name: 'emma' },
-];
-module.exports = { products, people };
+]
+module.exports = { products, people }
