+++ 02-express-tutorial/methods-public/index.html
@@ -19,7 +19,7 @@
       </div>
     </nav>
     <main>
-      <form action="/api/v1/people" method="POST">
+      <form action="/login" method="POST">
         <h3>Traditional Form</h3>
         <div class="form-row">
           <label for="name"> enter name </label>
