+++ 02-express-tutorial/methods-public/javascript.html
@@ -48,7 +48,7 @@
 
       const fetchPeople = async () => {
         try {
-          const { data } = await axios.get('/api/v1/people')
+          const { data } = await axios.get('/api/people')
 
           const people = data.data.map((person) => {
             return `<h5>${person.name}</h5>`
@@ -68,13 +68,13 @@
         const nameValue = input.value
 
         try {
-          const { data } = await axios.post('/api/v1/people', { name: nameValue })
+          const { data } = await axios.post('/api/people', { name: nameValue })
           const h5 = document.createElement('h5')
           h5.textContent = data.person
           result.appendChild(h5)
         } catch (error) {
           // console.log(error.response)
-          formAlert.textContent = error.response.data.message
+          formAlert.textContent = error.response.data.msg
         }
         input.value = ''
       })
